requirements.txt
channels==4.3.2
daphne==4.2.1
Django==6.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
twitchio==3.1.0
websockets==15.0.1

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.1
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
frozenlist==1.8.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
propcache==0.4.1
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
service-identity==24.2.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.2
yarl==1.22.0
zope.interface==8.1.1